% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents in a given year and state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Numeric state code as integer or character.}

\item{year}{Year as integer or character.}
}
\value{
\code{NULL}, but still plots the map.
}
\description{
Plots the geographical locations of reported accidents in a given year and U.S. state.
}
\examples{
\dontrun{
  fars_map_state(state.num = 1, year = 2013)
}

}
