% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read local csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to the target file.}
}
\value{
A tibble.
}
\description{
Reads a CSV file from a file path and returns it as a tibble
}
\examples{
\dontrun{
data <- fars_read("accident_2013.csv.bz2")
}

}
