% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read accident data files for given years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years of interest.}
}
\value{
A list of tibbles.
}
\description{
Reads accident data files for the specified years and returns the
\code{MONTH} and \code{year} columns for each successfully read file.
}
\details{
\code{NULL} is returned for non-existent elements.
}
\examples{
\dontrun{
file_list <- fars_read_years(c("2013", "2014", "2016"))
}

}
