% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Read in and summarize accident data files for given years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years of interest.}
}
\value{
A tibble with monthly accident counts per year.
}
\description{
Reads accident data files for the specified years and summarizes them per \code{year} and \code{MONTH}.
}
\examples{
\dontrun{
summary_tbl <- fars_summarize_years(c("2013", "2014", "2016"))
}

}
