% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create file name from year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year as integer or character.}
}
\value{
A character string with the file name.
}
\description{
Creates the file name of a compressed CSV file containing accident data
for a given year.
}
\examples{
\dontrun{
file_name <- make_filename(2015)
}

}
